+++ source/core/game.cpp
@@ -3,12 +3,37 @@
 #include <core/game.hpp>
 #include <entities/entity.hpp>
 
+#include <tools/editor.hpp>
+#include "tools/editorCam.cpp"
+
+#include "imgui.h"
+#include "backends/imgui_impl_sdl3.h"
+#include "backends/imgui_impl_sdlrenderer3.h"
+
 Game::Game() : running(true) {}
 Game::~Game() {}
 
 bool Game::init() {
 
+  // Initialize SDL and Graphics with our Renderer
   if (!renderer.init("Flicky", SCREEN_WIDTH, SCREEN_HEIGHT)) return false;
+
+  // ==== IMGUI SETUP / EDITOR SETUP ====
+  // Setup Dear ImGui context
+  IMGUI_CHECKVERSION();
+  ImGui::CreateContext();
+  ImGuiIO& io = ImGui::GetIO(); (void)io;
+  io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;     // Enable Keyboard Controls
+  io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;      // Enable Gamepad Controls
+
+  // Setup Dear ImGui style
+  ImGui::StyleColorsDark();
+
+  // Setup Platform/Renderer backends
+  ImGui_ImplSDL3_InitForSDLRenderer(renderer.getWindow(), renderer.getSDLRenderer());
+  ImGui_ImplSDLRenderer3_Init(renderer.getSDLRenderer());
+
+  // ==== GAME SETUP ====
 
   // Background Rendering
   std::string BGpath = std::string(PROJECT_ROOT) + "/assets/Sega Genesis 32X - Flicky - Area 1.png";
@@ -32,13 +57,26 @@ void Game::run() {
     last = now;
 
     processEvents();
+
+    // Start the Dear ImGui frame
+    ImGui_ImplSDLRenderer3_NewFrame();
+    ImGui_ImplSDL3_NewFrame();
+    ImGui::NewFrame();
+    
     update(dt);
+    
+    ImGui::Render();
     render();
   }
 }
 
 void Game::shutdown() {
   delete world;
+
+  ImGui_ImplSDLRenderer3_Shutdown();
+  ImGui_ImplSDL3_Shutdown();
+  ImGui::DestroyContext();
+
   renderer.shutdown();  
 
   SDL_Quit();
@@ -47,6 +85,7 @@ void Game::shutdown() {
 void Game::processEvents() {
   SDL_Event e;
   while (SDL_PollEvent(&e)) {
+    ImGui_ImplSDL3_ProcessEvent(&e);
     if (e.type == SDL_EVENT_QUIT) {
       running = false;
     }
@@ -56,13 +95,41 @@ void Game::processEvents() {
       if (e.key.key == SDLK_R) {
         world->loadLevel((std::string(PROJECT_ROOT) + "/levels/level1 - Copy.txt"));
       }
+
+      if (e.key.key == SDLK_F1) {
+        editor.toggle();  
+      }
+
+      if (e.key.key == SDLK_F2) {
+        if (pause == true) {
+          pause = false;
+          world->camera.follow(world->player);
+        }
+        else {
+          pause = true;
+          // Sync editor camera to current view center
+          edcam.x = world->camera.x + world->camera.width * 0.5f;
+          edcam.y = world->camera.y + world->camera.height * 0.5f;
+          
+          world->camera.follow(edcam.x, edcam.y);
+        }
+      }
+      
     }
+
   }
 
 }
 
 void Game::update(float dt) {
-  world->update(dt);
+  if (!pause) {
+    world->update(dt);
+  } else { 
+    edcam.handleInput(dt);
+    world->camera.follow(edcam.x, edcam.y);
+  }
+
+  editor.update(*world, renderer);
 }
 
 void Game::render() {
@@ -71,9 +138,11 @@ void Game::render() {
   if (!world->isRespawning()) {
     renderer.renderBackground(world->camera);
     renderer.renderWorld(*world);
-  } else {
+  } else {                                    // If player is hit / respawning the Screen turns to black
     renderer.drawBlackScreen();
   }
+
+  ImGui_ImplSDLRenderer3_RenderDrawData(ImGui::GetDrawData(), renderer.getSDLRenderer());
 
   renderer.present();
 }